Fixed RD-14975: Do not push LIMIT when OFFSET is specified

+++ src/multicorn.h
@@ -70,7 +70,7 @@ typedef struct MulticornPlanState
     int			startupCost;
     ConversionInfo **cinfos;
     List	   *pathkeys; /* list of MulticornDeparsedSortGroup) */
-    double		limit; /* store the LIMIT value, or -1 if no limit */
+    int64		limit; /* store the LIMIT value, or -1 if no limit */
     uint64		plan_id; /* unique identifier for this plan */
 
     /* For some reason, `baserel->reltarget->width` gets changed
@@ -93,7 +93,7 @@ typedef struct MulticornExecState
     List	   *qual_list;
     Datum	   *values;
     bool	   *nulls;
-    double	  	limit; /* store the LIMIT value, or -1 if no limit */
+    int64	  	limit; /* store the LIMIT value, or -1 if no limit */
     uint64	   	plan_id; /* unique identifier for this plan */
     ConversionInfo **cinfos;
     /* Common buffer to avoid repeated allocations */
